ML_model/requirements.txt
flask
scikit-learn
numpy
flask_cors
pandas
joblib
tensorflow


streamlit
google-generativeai
python-dotenv